+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Events building block API definition and schema to include event categories and sub categories
 - Separated non-pii data and pii data and database in profile building block
 - Profile building block rest api changed to Flask RESTFul api service
+- Profile building block non-pii interests has been changed by using category and subcategory
 
 ### Fixed
 - Events building block API definition to improve text.