+++ profileservice/CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [#73](https://github.com/rokwire/rokwire-building-blocks-api/issues/73)
 - profile building block rest api changed to Flask RESTFul api service
 [#31](https://github.com/rokwire/rokwire-building-blocks-api/issues/31)
+- non-pii interests has been modified with category and subcategory
+[#86](https://github.com/rokwire/rokwire-building-blocks-api/issues/86)
 
 ### Fixed